+++ bmc-fusionapps/src/main/java/com/oracle/bmc/fusionapps/FusionApplications.java
@@ -157,6 +157,19 @@ public interface FusionApplications extends AutoCloseable {
     CreateRefreshActivityResponse createRefreshActivity(CreateRefreshActivityRequest request);
 
     /**
+     * Attaches a service instance to the fusion pod.
+     *
+     * @param request The request object containing the details to send
+     * @return A response object containing details about the completed operation
+     * @throws BmcException when an error occurs.
+     * This operation uses RetryConfiguration.SDK_DEFAULT_RETRY_CONFIGURATION as default if no retry strategy is provided.
+     * The specifics of the default retry strategy are described here https://docs.oracle.com/en-us/iaas/Content/API/SDKDocs/javasdkconcepts.htm#javasdkconcepts_topic_Retries
+     *
+     * <b>Example: </b>Click <a href="https://docs.cloud.oracle.com/en-us/iaas/tools/java-sdk-examples/latest/fusionapps/CreateServiceAttachmentExample.java.html" target="_blank" rel="noopener noreferrer" >here</a> to see how to use CreateServiceAttachment API.
+     */
+    CreateServiceAttachmentResponse createServiceAttachment(CreateServiceAttachmentRequest request);
+
+    /**
      * Deletes the Fusion environment identified by it's OCID.
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
@@ -205,6 +218,18 @@ public interface FusionApplications extends AutoCloseable {
      * <b>Example: </b>Click <a href="https://docs.cloud.oracle.com/en-us/iaas/tools/java-sdk-examples/latest/fusionapps/DeleteRefreshActivityExample.java.html" target="_blank" rel="noopener noreferrer" >here</a> to see how to use DeleteRefreshActivity API.
      */
     DeleteRefreshActivityResponse deleteRefreshActivity(DeleteRefreshActivityRequest request);
+
+    /**
+     * Delete a service attachment by identifier
+     * @param request The request object containing the details to send
+     * @return A response object containing details about the completed operation
+     * @throws BmcException when an error occurs.
+     * This operation uses RetryConfiguration.SDK_DEFAULT_RETRY_CONFIGURATION as default if no retry strategy is provided.
+     * The specifics of the default retry strategy are described here https://docs.oracle.com/en-us/iaas/Content/API/SDKDocs/javasdkconcepts.htm#javasdkconcepts_topic_Retries
+     *
+     * <b>Example: </b>Click <a href="https://docs.cloud.oracle.com/en-us/iaas/tools/java-sdk-examples/latest/fusionapps/DeleteServiceAttachmentExample.java.html" target="_blank" rel="noopener noreferrer" >here</a> to see how to use DeleteServiceAttachment API.
+     */
+    DeleteServiceAttachmentResponse deleteServiceAttachment(DeleteServiceAttachmentRequest request);
 
     /**
      * Gets a DataMaskingActivity by identifier
@@ -525,6 +550,19 @@ public interface FusionApplications extends AutoCloseable {
      * <b>Example: </b>Click <a href="https://docs.cloud.oracle.com/en-us/iaas/tools/java-sdk-examples/latest/fusionapps/UpdateRefreshActivityExample.java.html" target="_blank" rel="noopener noreferrer" >here</a> to see how to use UpdateRefreshActivity API.
      */
     UpdateRefreshActivityResponse updateRefreshActivity(UpdateRefreshActivityRequest request);
+
+    /**
+     * Verify whether a service instance can be attached to the fusion pod
+     *
+     * @param request The request object containing the details to send
+     * @return A response object containing details about the completed operation
+     * @throws BmcException when an error occurs.
+     * This operation uses RetryConfiguration.SDK_DEFAULT_RETRY_CONFIGURATION as default if no retry strategy is provided.
+     * The specifics of the default retry strategy are described here https://docs.oracle.com/en-us/iaas/Content/API/SDKDocs/javasdkconcepts.htm#javasdkconcepts_topic_Retries
+     *
+     * <b>Example: </b>Click <a href="https://docs.cloud.oracle.com/en-us/iaas/tools/java-sdk-examples/latest/fusionapps/VerifyServiceAttachmentExample.java.html" target="_blank" rel="noopener noreferrer" >here</a> to see how to use VerifyServiceAttachment API.
+     */
+    VerifyServiceAttachmentResponse verifyServiceAttachment(VerifyServiceAttachmentRequest request);
 
     /**
      * Gets the pre-configured waiters available for resources for this service.